Refactor GUI into modular layout
